Move auth controllers to the user module
